issue#64 Fix issue with idling resources
reference: https://github.com/JakeWharton/okhttp-idling-resource/pull/16
Espresso expects callback when a resource is released. However since our OkHttp Dispatchers were reused and 2 compiting IdlingResource was registered, one overwrote the other's callback and this crashed the Espresso tests.
To resolve this I made sure when one is registeres, it's respects the current callback and wraps around it instead of overwriting it.